finished updated listings
