+++ src/Chapter02/Listing02.15.StringInterpolation.cs
@@ -1,4 +1,4 @@
-﻿namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_01
+﻿namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_15
 {
     public class Program
     {
