+++ src/Chapter02/Listing02.17.CombiningStringLiteralsWithInterpolation.cs
@@ -1,4 +1,4 @@
-﻿namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_01
+﻿namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_17
 {
     public class Program
     {
