--- src/Chapter02/Listing02.21.ExplicitCastExample.cs
@@ -1,13 +1,0 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_21
-{
-    public class Program
-    {
-        public static void Main()
-        {
-            #region INCLUDE
-            long longNumber = 50918309109;
-            int intNumber = (int)longNumber;
-            #endregion INCLUDE
-        }
-    }
-}
