+++ src/Chapter02/Listing02.21.UsingStaticDirectives.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_16
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_21
 {
     #region INCLUDE
     // The using directives allow you to drop the namespace
