+++ src/Chapter02/Listing02.22.UsingStringsLengthMember.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_17
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_22
 {
     #region INCLUDE
     public class PalindromeLength
