+++ src/Chapter02/Listing02.25.AssigningNullToAString.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_20
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_25
 {
     public class Program
     {
