+++ src/Chapter02/Listing02.27.OverflowingAnIntegerValue.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_22
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_27
 {
     public class Program
     {
