+++ src/Chapter02/Listing02.28.ACheckedBlockExample.cs
@@ -1,17 +1,21 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_24
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_28
 {
     public class Program
     {
         public static void Main()
         {
             #region INCLUDE
-            unchecked
+            #region HIGHLIGHT
+            checked
             {
+            #endregion HIGHLIGHT
                 // int.MaxValue equals 2147483647
                 int n = int.MaxValue;
                 n = n + 1;
                 Console.WriteLine(n);
+            #region HIGHLIGHT
             }
+            #endregion HIGHLIGHT
             #endregion INCLUDE
         }
     }
