+++ src/Chapter02/Listing02.29.AnUncheckedBlockExample.cs
@@ -5,10 +5,13 @@ namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_29
         public static void Main()
         {
             #region INCLUDE
-            bool boolean = true;
-            string text = boolean.ToString();
-            // Display "True"
-            Console.WriteLine(text);
+            unchecked
+            {
+                // int.MaxValue equals 2147483647
+                int n = int.MaxValue;
+                n = n + 1;
+                Console.WriteLine(n);
+            }
             #endregion INCLUDE
         }
     }
