+++ src/Chapter02/Listing02.30.NotUsingTheCastOperatorForAnImplicitCast.cs
@@ -1,0 +1,13 @@
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_30
+{
+    public class Program
+    {
+        public static void Main()
+        {
+            #region INCLUDE
+            int intNumber = 31416;
+            long longNumber = intNumber;
+            #endregion INCLUDE
+        }
+    }
+}
