+++ src/Chapter02/Listing02.32.UsingIntParseToConvertAStringToANumericDataType.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_27
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_32
 {
     public class Program
     {
