+++ src/Chapter02/Listing02.33.TypeConversionUsingSystemConvert.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_28
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_33
 {
     public class Program
     {
