+++ src/Chapter02/Listing02.34.UsingToStringToConvertToAString.cs
@@ -1,0 +1,15 @@
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_34
+{
+    public class Program
+    {
+        public static void Main()
+        {
+            #region INCLUDE
+            bool boolean = true;
+            string text = boolean.ToString();
+            // Display "True"
+            Console.WriteLine(text);
+            #endregion INCLUDE
+        }
+    }
+}
