+++ src/Chapter02/Listing02.36.UsingTryParseWithInlineOut.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_31
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter02.Listing02_36
 {
     public class Program
     {
